backend/requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docopt==0.6.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Sessions==0.1.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==22.0
pipreqs==0.4.11
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
yarg==0.1.9
